Users must explicitly require RedisStore monkeypatch.
