+++ spec/spec_helper.rb
@@ -14,3 +14,7 @@ ActiveJob::Base.queue_adapter = :test
 RSpec.configure do |c|
   c.include Helpers
 end
+
+if defined? ActiveSupport::Cache::RedisStore
+  require "active_job_status/redis"
+end
